farm_assistant/requirements.txt
dxcam>=0.0.5; platform_system=="Windows"
mss>=9.0.1
opencv-python>=4.8.0
numpy>=1.24
pytesseract>=0.3.10
pandas>=2.0
colorama>=0.4.6
typer>=0.9.0
pyyaml>=6.0
rapidfuzz>=3.0
